backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
alembic
geoalchemy2
python-jose[cryptography]
google-auth
httpx
python-multipart
python-dotenv
Mako
requests
